requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
jsonschema==4.21.1
numpy==1.25.2
pandas==2.0.3
uuid==1.30
google-generativeai==0.5.2
